introducing Board class
